[update] update experiment environment

+++ example/src/main/java/com/flucta/example/shortestpath/FloodingValue.java
@@ -2,35 +2,35 @@ package com.flucta.example.shortestpath;
 
 import com.flucta.core.common.Computable;
 
-public class FloodingValue implements Computable<Integer> {
-    private int value;
+public class FloodingValue implements Computable<Double> {
+    private double value;
 
-    public FloodingValue(int value) {
+    public FloodingValue(double value) {
         this.value = value;
     }
 
     @Override
-    public Computable<Integer> add(Computable<Integer> computable) {
+    public Computable<Double> add(Computable<Double> computable) {
         return new FloodingValue(this.getValue() + computable.getValue());
     }
 
     @Override
-    public FloodingValue minus(Computable<Integer> computable) {
+    public FloodingValue minus(Computable<Double> computable) {
         return new FloodingValue(this.getValue() - computable.getValue());
     }
 
     @Override
-    public FloodingValue multiply(Computable<Integer> computable) {
+    public FloodingValue multiply(Computable<Double> computable) {
         return new FloodingValue(this.getValue() * computable.getValue());
     }
 
     @Override
-    public Integer getValue() {
+    public Double getValue() {
         return this.value;
     }
 
     @Override
-    public void setValue(Integer value) {
+    public void setValue(Double value) {
         this.value = value;
     }
 }
